Extract utils from app
